+++ app/src/main/java/app/revanced/integrations/twitter/settings/featureflags/CustomAdapter.java
@@ -1,14 +1,17 @@
 package app.revanced.integrations.twitter.settings.featureflags;
 
 import android.annotation.SuppressLint;
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
-import android.widget.CompoundButton;
 import android.widget.Switch;
 import android.widget.TextView;
 import androidx.recyclerview.widget.RecyclerView;
 import app.revanced.integrations.shared.Utils;
 
+import java.lang.reflect.InvocationTargetException;
+import java.lang.reflect.Method;
+import java.lang.reflect.Modifier;
 import java.util.ArrayList;
 
 @SuppressWarnings("unused")
@@ -33,22 +36,59 @@ public class CustomAdapter extends RecyclerView.e<CustomAdapter.ViewHolder> {
         this.itemCheckedChangeListener = itemCheckedChangeListener;
     }
 
+    public void notifyItemChanged(int i) {
+        Method notifyMethod = null;
+        for (Method declaredMethod : this.getClass().getSuperclass().getDeclaredMethods()) {
+            if (declaredMethod.getParameterTypes().length==1) {
+                int modifier = declaredMethod.getModifiers();
+                Class<?> parameterType = declaredMethod.getParameterTypes()[0];
+                if (parameterType==Integer.TYPE && (modifier & Modifier.FINAL) == Modifier.FINAL) {
+                    notifyMethod = declaredMethod;
+                }
+            }
+        }
+
+        try {
+            if (notifyMethod!=null) {
+                notifyMethod.invoke(this, i);
+            }
+        } catch (IllegalAccessException | InvocationTargetException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
     public final class ViewHolder extends RecyclerView.c0 {
+        private int getAdapterPosition() {
+            Method getAdapterPositionMethod = null;
+            for (Method declaredMethod : this.getClass().getSuperclass().getDeclaredMethods()) {
+                if (declaredMethod.getReturnType()==Integer.TYPE && declaredMethod.getParameterTypes().length==0) {
+                    getAdapterPositionMethod = declaredMethod;
+                    break;
+                }
+            }
+
+            if (getAdapterPositionMethod!=null) {
+                try {
+                    return (int) getAdapterPositionMethod.invoke(this);
+                } catch (IllegalAccessException | InvocationTargetException e) {
+                    throw new RuntimeException(e);
+                }
+            }
+            return -1;
+        }
+
         // a0() = getAdapterPosition()
         public ViewHolder(CustomAdapter adapter, View view) {
             super(view);
             flagTextView = view.findViewById(Utils.getResourceIdentifier("textView", "id"));
 
             view.setOnClickListener(view1 -> {
-                if (itemClickListener!=null) itemClickListener.onClick(a0());
+                if (itemClickListener!=null) itemClickListener.onClick(getAdapterPosition());
             });
 
             enabled = view.findViewById(Utils.getResourceIdentifier("enabled", "id"));
-            enabled.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
-                @Override
-                public void onCheckedChanged(CompoundButton compoundButton, boolean b) {
-                    if (itemCheckedChangeListener!=null) itemCheckedChangeListener.onCheck(b, a0());
-                }
+            enabled.setOnCheckedChangeListener((compoundButton, b) -> {
+                if (itemCheckedChangeListener!=null) itemCheckedChangeListener.onCheck(b, getAdapterPosition());
             });
         }
     }